Refactoring contact component, splitting to separate ones
