+++ types/index.ts
@@ -2,3 +2,4 @@ export * from "./sanity-base";
 export * from "./sanity-content";
 export * from "./sanity-theme";
 export * from "./props";
+export * from "./form";
