+++ types/props.ts
@@ -67,3 +67,15 @@ export interface ContactFormProps {
   theme: Theme | null;
   about: About | null;
 }
+
+export interface FormFieldProps {
+  id: string;
+  name: string;
+  type?: string;
+  value: string;
+  onChange: (
+    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
+  ) => void;
+  placeholder: string;
+  rows?: number;
+}
